// src/global/enums.ts
export enum COST_TYPE {
  HOURLY = 'HOURLY',
  FIXED_COST = 'FIXED_COST',
}
export enum ROLE {
  SERVICE_PROVIDER = 'SERVICE_PROVIDER',
  CLIENT = 'CLIENT',
  SERVICE_PROVIDER_COMPANY = 'SERVICE_PROVIDER_COMPANY',
  ADMIN = 'ADMIN'
}

export enum COMPANY_ROLE {
  ADMIN = 'ADMIN',
  TEAM_MEMBER = 'TEAM_MEMBER',
}

export enum GENDER {
  MALE = 'MALE',
  FEMALE = 'FEMALE'
}

export enum SERVICE_STATUS {
  DRAFT = 'DRAFT',
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE'
}

export enum COMMISSIONS_STATUS {
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE'
}

export enum SUPPORT_CATEGORY {
  PROJECT = 'PROJECT',
  OTHER = 'OTHER'
}

export enum SKILL_STATUS {
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE',
  DELETED = 'DELETED'
}

export enum SUB_CATEGORIES_STATUS {
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE',
  DELETED = 'DELETED'
}

export enum CATEGORY_STATUS {
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE',
  DELETED = 'DELETED'
}

export enum PROJECT_STATUS {
  DRAFT = 'DRAFT',
  POSTED = 'POSTED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE',
  CANCEL = 'CANCEL'
}

export enum PROJECT_PREFS {
  INDIVIDUAL = 'INDIVIDUAL',
  COMPANY = 'COMPANY',
  BOTH = 'BOTH'
}
export enum SLOT_STATUS {
  REMOVED = 'REMOVED',
  AVAILABLE = 'AVAILABLE',
  BOOKED = 'BOOKED'
}

export enum BOOKING_STATUS {
  VISITING_CHARGES_PENDING = 'VISITING_CHARGES_PENDING',
  SCHEDULED = 'SCHEDULED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  COMPLETED = 'COMPLETED',
  CANCELED = 'CANCELED'
}

export enum USER_STATUS {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  BLACKLIST = 'BLACKLIST',
  DELETED = 'DELETED'
}

export enum MILESTONE_STATUS {
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  PAYMENT_IN_PROGRESS = 'PAYMENT_IN_PROGRESS',
  PAYMENT_DONE = 'PAYMENT_DONE',
  PAYMENT_RELEASE = 'PAYMENT_RELEASE',
  PAYMENT_RELEASE_APPROVE = 'PAYMENT_RELEASE_APPROVE'
}

export enum LOGINTYPE {
  FACEBOOK = 'facebook',
  GOOGLE = 'google',
  PROTOCALL = 'protocall',
  apple = 'apple'
}

export enum CATEGORY_ASSOCIATE {
  OFFSITE = 'OFFSITE',
  ONSITE = 'ONSITE'
}

export enum COMMISSION_ROLE {
  SERVICE_PROVIDER = 'SERVICE_PROVIDER',
  SERVICE_PROVIDER_COMPANY = 'SERVICE_PROVIDER_COMPANY',
  ALL = 'ALL'
}

export enum PAYOUT_STATUS {
  TRANSFERRED = 'TRANSFERRED',
  PENDING = 'PENDING'
}

export enum CHAT_PROFILE_TYPE {
  CLI = 'CLI',
  PRO = 'PRO',
  COM = 'COM',
  ADM = 'ADM'
}

export enum TAX_TYPE {
  GENERAL = 'GENERAL',
  SPECIFIC = 'SPECIFIC'
}

export enum TAX_MASTER_STATUS {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE'
}

export enum CHAT_ENTITY_TYPE {
  PROJECT = 'PROJECT',
  BOOKING = 'BOOKING',
  SUPPORT = 'SUPPORT',
  CLIENT_SP = 'CLIENT_SP'

}